refactor: Altered animal_detector() algorithm logic
Also answer to question 6 was changed to include
testing on more images, an extra dog and a non
human or dog image. Both used as testing subjects
to handle the case of styled image of a dog and
one to handle the case of an image not having a
dog or human face in it.